@@ -1,59 +1,80 @@
 {
-  "name": "template",
+  "name": "nish1896-react-libs",
   "version": "1.0.0",
   "author": "Nishant Kohli",
-  "private": true,
   "dependencies": {
-    "@fontsource/covered-by-your-grace": "^4.5.6",
-    "@fontsource/crimson-pro": "^4.5.7",
-    "@fontsource/roboto": "^4.5.5",
+    "@bpmn-io/form-js-editor": "^1.7.0",
+    "@bpmn-io/form-js-viewer":"^1.7.0",
+    "@ckeditor/ckeditor5-build-classic": "^41.0.0",
+    "@ckeditor/ckeditor5-code-block": "^41.0.0",
+    "@ckeditor/ckeditor5-find-and-replace": "^41.0.0",
+    "@ckeditor/ckeditor5-react": "^6.2.0",
+    "@emotion/react": "^11.11.3",
+    "@emotion/styled": "^11.11.0",
+    "@faker-js/faker": "^8.4.0",
+    "@fontsource/covered-by-your-grace": "^5.0.8",
+    "@fontsource/crimson-pro": "^5.0.16",
+    "@fontsource/roboto": "^5.0.8",
     "@hookform/resolvers": "^2.9.7",
-    "@material-ui/core": "^4.11.2",
-    "@material-ui/icons": "^4.11.2",
-    "@material-ui/lab": "^4.0.0-alpha.61",
-    "@reduxjs/toolkit": "^1.6.1",
+    "@mui/icons-material": "^5.15.7",
+    "@mui/material": "^5.15.7",
+    "@mui/styles": "^5.15.7",
+    "@reduxjs/toolkit": "^2.1.0",
     "@tailwindcss/postcss7-compat": "^2.0.2",
-    "@types/jest": "^28.1.6",
-    "@types/node": "^18.7.1",
-    "@types/react": "^18.0.17",
-    "@types/react-dom": "^18.0.6",
     "@zoomus/websdk": "^1.9.6",
-    "algoliasearch": "^4.18.0",
+    "algoliasearch": "^4.22.1",
     "aos": "^2.3.4",
     "autoprefixer": "9",
-    "axios": "^0.21.1",
+    "axios": "^1.6.7",
     "bootstrap": "^5.3.2",
     "classnames": "^2.3.2",
     "cogo-toast": "^4.2.3",
     "customize-cra": "^1.0.0",
-    "faker": "^5.5.3",
     "final-form": "^4.20.7",
-    "formik": "^2.2.6",
+    "formik": "^2.4.5",
     "jquery": "^3.5.1",
+    "mui-rte": "^2.0.1",
     "postcss": "7",
-    "react": "^17.0.1",
-    "react-beautiful-dnd": "^13.1.0",
+    "react": "^18.0.2",
+    "react-beautiful-dnd": "^13.1.1",
     "react-charts": "^2.0.0-beta.7",
-    "react-csv": "^2.0.3",
-    "react-dom": "^17.0.1",
+    "react-color": "^2.19.3",
+    "react-csv": "^2.2.2",
+    "react-dom": "^18.0.2",
     "react-final-form": "^6.5.9",
-    "react-hook-form": "^7.34.0",
-    "react-icons": "^4.2.0",
-    "react-instantsearch": "^7.0.1",
-    "react-player": "^2.9.0",
-    "react-query": "^3.17.2",
-    "react-redux": "^7.2.5",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.1",
-    "react-toastify": "^7.0.4",
-    "react-window": "^1.8.6",
+    "react-hook-form": "^7.50.1",
+    "react-icons": "^5.0.1",
+    "react-instantsearch": "^7.5.5",
+    "react-player": "^2.14.1",
+    "react-query": "^3.39.3",
+    "react-redux": "^9.1.0",
+    "react-router-dom": "^6.22.0",
+    "react-scripts": "5.0.1",
+    "react-toastify": "^10.0.4",
+    "react-window": "^1.8.10",
     "redux-persist": "^6.0.0",
+    "short-unique-id": "^5.0.3",
     "tailwindcss": "npm:@tailwindcss/postcss7-compat",
-    "typescript": "^4.7.4",
     "video.js": "^7.10.2",
     "videojs-youtube": "^2.6.1",
     "web-vitals": "^0.2.4",
     "yup": "^0.32.11"
+  },
+  "devDependencies": {
+    "@nish1896/eslint-config": "^1.0.3",
+    "@types/jest": "^28.1.6",
+    "@types/node": "^18.7.1",
+    "@types/react": "^18.0.17",
+    "@types/react-dom": "^18.0.6",
+    "@types/react-router-dom": "^5.3.3",
+    "cypress": "^12.5.1",
+    "eslint": "^8.56.0",
+    "husky": "^9.0.10",
+    "lint-staged": "^10.2.13",
+    "prettier": "^2.2.1",
+    "react-app-rewired": "^2.1.8",
+    "style-dictionary": "^3.9.2",
+    "typescript": "^5.3.3"
   },
   "scripts": {
     "build-dictionary": "style-dictionary build --config ./style-dictionary/config.js",
