@@ -1,9 +1,9 @@
 import { Fragment, useState, useEffect } from 'react';
-import TextField from '@material-ui/core/TextField';
-import { default as MUIAutocomplete } from '@material-ui/lab/Autocomplete';
-import CircularProgress from '@material-ui/core/CircularProgress';
-import IconButton from '@material-ui/core/IconButton';
-import CloseIcon from '@material-ui/icons/Close';
+import TextField from '@mui/material/TextField';
+import { default as MUIAutocomplete } from '@mui/material/Autocomplete';
+import CircularProgress from '@mui/material/CircularProgress';
+import IconButton from '@mui/material/IconButton';
+import CloseIcon from '@mui/icons-material/Close';
 import { fetchAlgoliaData } from './algolia.service';
 import { ProductInfo } from '../types';
 
@@ -20,7 +20,7 @@ const Autocomplete = () => {
      *  function will only call once in every 500 ms.
      */
     const fetchProducts = setTimeout(async () => {
-      if (Boolean(inputValue)) {
+      if (inputValue) {
         setLoading(true);
         const products = await fetchAlgoliaData(inputValue);
         setOptions(products ?? []);
@@ -48,32 +48,35 @@ const Autocomplete = () => {
         setOpen(false);
       }}
       value={value}
-      filterOptions={(x) => x}
+      filterOptions={x => x}
       noOptionsText={
-        Boolean(inputValue)
+        inputValue
           ? 'No Results Found'
           : 'Type something to fetch results...'
       }
-      onChange={(event: any, newValue: ProductInfo | null) => {
+      onChange={(_, newValue: ProductInfo | null) => {
         setOptions(newValue ? [newValue, ...options] : options);
         setValue(newValue);
       }}
-      onInputChange={(event, newInputValue) => {
+      onInputChange={(_, newInputValue) => {
         setInputValue(newInputValue);
       }}
-      getOptionLabel={(option) => option.name}
+      getOptionLabel={option => option.name}
       options={options}
       loading={loading}
-      renderInput={(params) => (
+      renderInput={params => (
         <TextField
           {...params}
           value={inputValue}
           fullWidth
-          variant='outlined'
+          variant="outlined"
           placeholder="Search for products in Autocomplete...."
           InputProps={{
             ...params.InputProps,
-            style: { padding: '5px 10px', background: 'white' },
+            style: {
+              padding: '5px 10px',
+              background: 'white',
+            },
             endAdornment: (
               <Fragment>
                 {loading ? (
