@@ -5,8 +5,8 @@ import {
   ClearRefinements,
 } from 'react-instantsearch';
 import classNames from 'classnames';
-import IconButton from '@material-ui/core/IconButton';
-import CloseIcon from '@material-ui/icons/Close';
+import IconButton from '@mui/material/IconButton';
+import CloseIcon from '@mui/icons-material/Close';
 import AlgoliaConfig from '../algolia-config';
 
 /**
@@ -17,22 +17,22 @@ import AlgoliaConfig from '../algolia-config';
  */
 
 const CurrentRefinementWidget = (
-  props: CurrentRefinementsProps
+  props: CurrentRefinementsProps,
 ): ReactElement => {
   const { items, refine } = useCurrentRefinements(props);
   return (
     <div
-      className={classNames('ais-CurrentRefinements--root', {
-        'ais-CurrentRefinements--noRefinementRoot': items.length === 0,
-      })}
+      className={classNames('ais-CurrentRefinements--root', { 'ais-CurrentRefinements--noRefinementRoot': items.length === 0 })}
     >
       {items.length > 0 && <h6>Filters Applied</h6>}
-      {items.map((refinementItem) => {
-        const isPriceFilter =
-          refinementItem.attribute === AlgoliaConfig.FACET_ATTRIBUTES.price;
+      {items.map(refinementItem => {
+        const isPriceFilter
+          = refinementItem.attribute === AlgoliaConfig.FACET_ATTRIBUTES.price;
         return (
           <Fragment key={refinementItem.attribute}>
-            <h5>{refinementItem.attribute.toUpperCase()}</h5>
+            <h5>
+              {refinementItem.attribute.toUpperCase()}
+            </h5>
             {isPriceFilter ? (
               <div
                 key={refinementItem.label}
@@ -42,12 +42,12 @@ const CurrentRefinementWidget = (
                   {refinementItem?.refinements?.[0]?.value}
                   {` <= ${AlgoliaConfig.FACET_ATTRIBUTES.price} < `}
                   {Math.round(
-                    parseFloat(`${refinementItem?.refinements?.[1]?.value}`)
+                    parseFloat(`${refinementItem?.refinements?.[1]?.value}`),
                   )}
                 </span>
                 <IconButton
                   onClick={() => {
-                    refinementItem.refinements.map((ri) => refine(ri));
+                    refinementItem.refinements.map(ri => refine(ri));
                   }}
                 >
                   <CloseIcon />
@@ -55,13 +55,17 @@ const CurrentRefinementWidget = (
               </div>
             ) : (
               <Fragment>
-                {refinementItem.refinements.map((refinement) => (
+                {refinementItem.refinements.map(refinement => (
                   <div
                     key={refinement.label}
                     className="ais-CurrentRefinements--refinement-item"
                   >
-                    <span>{refinement.label}</span>
-                    <span>{` (${refinement.count})`}</span>
+                    <span>
+                      {refinement.label}
+                    </span>
+                    <span>
+                      {` (${refinement.count})`}
+                    </span>
                     <IconButton onClick={() => refine(refinement)}>
                       <CloseIcon />
                     </IconButton>
@@ -78,9 +82,7 @@ const CurrentRefinementWidget = (
           button: 'ais-ClearRefinements--button',
           disabledButton: 'ais-ClearRefinements--button-disabled',
         }}
-        translations={{
-          resetButtonText:'Clear Refinements',
-        }}
+        translations={{ resetButtonText: 'Clear Refinements' }}
         // excludedAttributes={['brand']}
       />
     </div>
