@@ -1,6 +1,8 @@
+/* eslint-disable @typescript-eslint/ban-ts-comment */
+
 import { useState } from 'react';
 import { useRange, UseRangeProps } from 'react-instantsearch';
-import Slider from '@material-ui/core/Slider';
+import Slider from '@mui/material/Slider';
 
 const RangeSlider = (props: UseRangeProps) => {
   const { start, range, refine } = useRange(props);
@@ -12,13 +14,18 @@ const RangeSlider = (props: UseRangeProps) => {
 
   const [value, setValue] = useState<number[]>([from, to]);
 
-  const handleChange = (event: Event, newRange: number | number[]) => {
+  const handleChange = (_: Event, newRange: number | number[]) => {
     setValue(newRange as number[]);
     // @ts-ignore
     refine([newRange?.[0], newRange?.[1]]);
   };
   return (
-    <div style={{ padding: 10, display: 'flex', alignItems: "center"}}>
+    <div style={{
+      padding: 10,
+      display: 'flex',
+      alignItems: 'center',
+    }}
+    >
       <span>0</span>
       <Slider
         min={props.min}
