@@ -1,7 +1,7 @@
 import { ReactElement } from 'react';
 import { useRefinementList, RefinementListProps } from 'react-instantsearch';
 import classNames from 'classnames';
-import TextField from '@material-ui/core/TextField';
+import TextField from '@mui/material/TextField';
 
 const ListResults = (props: RefinementListProps): ReactElement => {
   /*
@@ -15,16 +15,14 @@ const ListResults = (props: RefinementListProps): ReactElement => {
     <div className="ais-RefinementList">
       <TextField
         placeholder="Search for brands"
-        style={{ margin: '10px 0px' }}
+        sx={{ margin: '10px 0px' }}
         inputProps={{ style: { padding: '8px 12px' } }}
-        onChange={(e) => searchForItems(e.currentTarget.value)}
+        onChange={e => searchForItems(e.currentTarget.value)}
       />
-      {items.map((item) => (
+      {items.map(item => (
         <div
           key={item.value}
-          className={classNames('ais-RefinementList-item', {
-            'ais-RefinementList-item--selected': item.isRefined,
-          })}
+          className={classNames('ais-RefinementList-item', { 'ais-RefinementList-item--selected': item.isRefined })}
         >
           <label className="ais-RefinementList-label">
             <input
@@ -34,7 +32,9 @@ const ListResults = (props: RefinementListProps): ReactElement => {
               checked={item.isRefined}
               onChange={() => refine(item.value)}
             />
-            <span className="ais-RefinementList-labelText">{item.label}</span>
+            <span className="ais-RefinementList-labelText">
+              {item.label}
+            </span>
             <span className="ais-RefinementList-count">
               {` (${item.count})`}
             </span>
