@@ -1,9 +1,9 @@
 import AosDiv from '.';
 
 it('matches AosDiv Text', () => {
-    cy.mount(<AosDiv />);
-    cy.get('[data-aos="fade-right"]').should(
-        'contains.text',
-        'some scrollable text animated'
-    );
+  cy.mount(<AosDiv />);
+  cy.get('[data-aos="fade-right"]').should(
+    'contains.text',
+    'some scrollable text animated',
+  );
 });
