@@ -2,104 +2,100 @@ import React from 'react';
 import { DragDropContext, Droppable, Draggable } from 'react-beautiful-dnd';
 
 // fake data generator
-const getItems = (count) =>
-    Array.from({ length: count }, (v, k) => k).map((k, index) => ({
-        id: `item-${k}`,
-        content: `item ${k}`,
-        position: index,
-    }));
+const getItems = count =>
+  Array.from({ length: count }, (v, k) => k).map((k, index) => ({
+    id: `item-${k}`,
+    content: `item ${k}`,
+    position: index,
+  }));
 
 // a little function to help us with reordering the result
 const reorder = (list, startIndex, endIndex) => {
-    const result = Array.from(list);
-    const [removed] = result.splice(startIndex, 1);
-    result.splice(endIndex, 0, removed);
+  const result = Array.from(list);
+  const [removed] = result.splice(startIndex, 1);
+  result.splice(endIndex, 0, removed);
 
-    /* Updating new positions */
-    result.map((ele, index) => (ele.position = index));
-    return result;
+  /* Updating new positions */
+  result.map((ele, index) => (ele.position = index));
+  return result;
 };
 
 const grid = 8;
 
 const getItemStyle = (isDragging, draggableStyle) => ({
-    // some basic styles to make the items look a bit nicer
-    userSelect: 'none',
-    padding: grid * 2,
-    margin: `0 0 ${grid}px 0`,
-    // change background colour if dragging
-    background: isDragging ? 'lightgreen' : 'grey',
+  // some basic styles to make the items look a bit nicer
+  userSelect: 'none',
+  padding: grid * 2,
+  margin: `0 0 ${grid}px 0`,
+  // change background colour if dragging
+  background: isDragging ? 'lightgreen' : 'grey',
 
-    // styles we need to apply on draggables
-    ...draggableStyle,
+  // styles we need to apply on draggables
+  ...draggableStyle,
 });
 
-const getListStyle = (isDraggingOver) => ({
-    background: isDraggingOver ? 'lightblue' : 'lightgrey',
-    height: 400,
-    padding: grid,
-    width: 450,
-    overflow: 'scroll',
+const getListStyle = isDraggingOver => ({
+  background: isDraggingOver ? 'lightblue' : 'lightgrey',
+  height: 400,
+  padding: grid,
+  width: 450,
+  overflow: 'scroll',
 });
 
 const DragNDrop = () => {
-    const [items, setItems] = React.useState(getItems(15));
+  const [items, setItems] = React.useState(getItems(15));
 
-    const onDragEnd = (result) => {
-        // dropped outside the list
-        if (!result.destination) {
-            return;
-        }
+  const onDragEnd = result => {
+    // dropped outside the list
+    if (!result.destination) {
+      return;
+    }
 
-        const orderedItems = reorder(
-            items,
-            result.source.index,
-            result.destination.index
-        );
-        console.log(result.source.index, result.destination.index);
-        console.log('orderedItems: ', orderedItems);
-
-        setItems(orderedItems);
-    };
-
-    return (
-        <div className="fullscreen">
-            <DragDropContext onDragEnd={(res) => onDragEnd(res)}>
-                <Droppable droppableId="droppable">
-                    {(provided, snapshot) => (
-                        <div
-                            {...provided.droppableProps}
-                            ref={provided.innerRef}
-                            style={getListStyle(snapshot.isDraggingOver)}
-                        >
-                            {items.map((item, index) => (
-                                <Draggable
-                                    key={item.id}
-                                    draggableId={item.id}
-                                    index={index}
-                                >
-                                    {(provided, snapshot) => (
-                                        <div
-                                            ref={provided.innerRef}
-                                            {...provided.draggableProps}
-                                            {...provided.dragHandleProps}
-                                            style={getItemStyle(
-                                                snapshot.isDragging,
-                                                provided.draggableProps.style
-                                            )}
-                                        >
-                                            {item.content}
-                                        </div>
-                                    )}
-                                </Draggable>
-                            ))}
-                            {provided.placeholder}
-                        </div>
-                    )}
-                </Droppable>
-            </DragDropContext>
-        </div>
+    const orderedItems = reorder(
+      items,
+      result.source.index,
+      result.destination.index,
     );
+    console.log(result.source.index, result.destination.index);
+    console.log('orderedItems: ', orderedItems);
+
+    setItems(orderedItems);
+  };
+
+  return (
+    <div className="fullscreen">
+      <DragDropContext onDragEnd={res => onDragEnd(res)}>
+        <Droppable droppableId="droppable">
+          {(provided, snapshot) => (
+            <div
+              {...provided.droppableProps}
+              ref={provided.innerRef}
+              style={getListStyle(snapshot.isDraggingOver)}
+            >
+              {items.map((item, index) => (
+                <Draggable key={item.id} draggableId={item.id} index={index}>
+                  {(provided, snapshot) => (
+                    <div
+                      ref={provided.innerRef}
+                      {...provided.draggableProps}
+                      {...provided.dragHandleProps}
+                      style={getItemStyle(
+                        snapshot.isDragging,
+                        provided.draggableProps.style,
+                      )}
+                    >
+                      {item.content}
+                    </div>
+                  )}
+                </Draggable>
+              ))}
+              {provided.placeholder}
+            </div>
+          )}
+        </Droppable>
+      </DragDropContext>
+    </div>
+  );
 };
 
 export default DragNDrop;
