@@ -1,8 +1,8 @@
 import FinalForm from '.';
 
 describe('<FinalForm />', () => {
-    it('renders', () => {
-        // see: https://on.cypress.io/mounting-react
-        cy.mount(<FinalForm />);
-    });
+  it('renders', () => {
+    // see: https://on.cypress.io/mounting-react
+    cy.mount(<FinalForm />);
+  });
 });
