@@ -1,20 +1,12 @@
 module.exports = {
-    color: {
-        base: {
-            primary: {
-                value: '#61dafb',
-            },
-            secondary: {
-                value: '#fc7aa5',
-            },
-        },
-        text: {
-            value: '#666666',
-        },
-        background: {
-            value: '#EEEEEE',
-        },
-        withOpacity: { value: '#0066991D' },
-        default: { value: '{color.base.primary.value}' },
+  color: {
+    base: {
+      primary: { value: '#61dafb' },
+      secondary: { value: '#fc7aa5' },
     },
+    text: { value: '#666666' },
+    background: { value: '#EEEEEE' },
+    withOpacity: { value: '#0066991D' },
+    default: { value: '{color.base.primary.value}' },
+  },
 };
